backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pydantic
pydantic-settings
rank-bm25
sentence-transformers
faiss-cpu
numpy
requests
websockets
python-multipart
SpeechRecognition
pydub
gtts
python-jose[cryptography]
passlib[bcrypt]
python-multipart
slowapi
pyttsx3>=2.90
httpx

